Refactor Minions
